Move acceptor plugin payload structs into separate files
